Ajustes na estrutura de pacotes. Os pacotes de service e specification vão para dentro de model
