+++ src/main/java/com/sistemaRegistroVerificacao/controller/SalaController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.sistemaRegistroVerificacao.exception.CampoInvalidoException;
 import com.sistemaRegistroVerificacao.model.entity.Sala;
-import com.sistemaRegistroVerificacao.seletor.SalaSeletor;
+import com.sistemaRegistroVerificacao.model.seletor.SalaSeletor;
 import com.sistemaRegistroVerificacao.service.SalaService;
 
 @RestController
