+++ src/main/java/com/sistemaRegistroVerificacao/model/seletor/AtividadeSeletor.java
@@ -1,0 +1,5 @@
+package com.sistemaRegistroVerificacao.model.seletor;
+
+public class AtividadeSeletor {
+    
+}
