+++ src/main/java/com/sistemaRegistroVerificacao/model/seletor/SalaSeletor.java
@@ -1,0 +1,5 @@
+package com.sistemaRegistroVerificacao.model.seletor;
+
+public class SalaSeletor {
+    
+}
