+++ src/main/java/com/sistemaRegistroVerificacao/model/seletor/ServicoPrestadoSeletor.java
@@ -1,0 +1,5 @@
+package com.sistemaRegistroVerificacao.model.seletor;
+
+public class ServicoPrestadoSeletor {
+    
+}
