+++ src/main/java/com/sistemaRegistroVerificacao/model/specification/AtividadeSpecification.java
@@ -1,9 +1,9 @@
-package com.sistemaRegistroVerificacao.specification;
+package com.sistemaRegistroVerificacao.model.specification;
 
 import org.springframework.data.jpa.domain.Specification;
 
 import com.sistemaRegistroVerificacao.model.entity.Atividade;
-import com.sistemaRegistroVerificacao.seletor.AtividadeSeletor;
+import com.sistemaRegistroVerificacao.model.seletor.AtividadeSeletor;
 
 public class AtividadeSpecification {
