+++ src/main/java/com/sistemaRegistroVerificacao/model/specification/NivelAcessoSpecification.java
@@ -1,9 +1,9 @@
-package com.sistemaRegistroVerificacao.specification;
+package com.sistemaRegistroVerificacao.model.specification;
 
 import org.springframework.data.jpa.domain.Specification;
 
 import com.sistemaRegistroVerificacao.model.entity.NivelAcesso;
-import com.sistemaRegistroVerificacao.seletor.NivelAcessoSeletor;
+import com.sistemaRegistroVerificacao.model.seletor.NivelAcessoSeletor;
 
 public class NivelAcessoSpecification {
