+++ src/main/java/com/sistemaRegistroVerificacao/model/specification/OcorrenciaSpecification.java
@@ -1,0 +1,5 @@
+package com.sistemaRegistroVerificacao.model.specification;
+
+public class OcorrenciaSpecification {
+    
+}
