+++ src/main/java/com/sistemaRegistroVerificacao/model/specification/SalaSpecification.java
@@ -1,9 +1,9 @@
-package com.sistemaRegistroVerificacao.specification;
+package com.sistemaRegistroVerificacao.model.specification;
 
 import org.springframework.data.jpa.domain.Specification;
 
 import com.sistemaRegistroVerificacao.model.entity.Sala;
-import com.sistemaRegistroVerificacao.seletor.SalaSeletor;
+import com.sistemaRegistroVerificacao.model.seletor.SalaSeletor;
 
 public class SalaSpecification {
