+++ src/main/java/com/sistemaRegistroVerificacao/model/specification/ServicoPrestadoSpecification.java
@@ -1,9 +1,9 @@
-package com.sistemaRegistroVerificacao.specification;
+package com.sistemaRegistroVerificacao.model.specification;
 
 import org.springframework.data.jpa.domain.Specification;
 
 import com.sistemaRegistroVerificacao.model.entity.ServicoPrestado;
-import com.sistemaRegistroVerificacao.seletor.ServicoPrestadoSeletor;
+import com.sistemaRegistroVerificacao.model.seletor.ServicoPrestadoSeletor;
 
 public class ServicoPrestadoSpecification {
