--- src/main/java/com/sistemaRegistroVerificacao/seletor/NivelAcessoSeletor.java
@@ -1,5 +1,0 @@
-package com.sistemaRegistroVerificacao.seletor;
-
-public class NivelAcessoSeletor {
-    
-}
