--- src/main/java/com/sistemaRegistroVerificacao/seletor/ServicoPrestadoSeletor.java
@@ -1,5 +1,0 @@
-package com.sistemaRegistroVerificacao.seletor;
-
-public class ServicoPrestadoSeletor {
-    
-}
