--- src/main/java/com/sistemaRegistroVerificacao/seletor/UsuarioSeletor.java
@@ -1,5 +1,0 @@
-package com.sistemaRegistroVerificacao.seletor;
-
-public class UsuarioSeletor {
-    
-}
