--- src/main/java/com/sistemaRegistroVerificacao/specification/OcorrenciaSpecification.java
@@ -1,5 +1,0 @@
-package com.sistemaRegistroVerificacao.specification;
-
-public class OcorrenciaSpecification {
-    
-}
